+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Add `FrontendUser.membershipNumber` (#894)
+
 ### Changed
 
 ### Deprecated